requirements.txt
torch>=2.0.0
transformers>=4.35.0
huggingface_hub>=0.20.0
python-dotenv>=1.0.0
modal-client>=0.50.2
diffusers>=0.24.0
accelerate
imageio
modal==1.1.1
transformers
